Rework Nuke Tests
