--- NukeExtensions/NukeExtensions.h
@@ -1,19 +1,0 @@
-//
-//  NukeExtensions.h
-//  NukeExtensions
-//
-//  Created by Alexander Grebenyuk on 6/12/22.
-//  Copyright © 2022 Alexander Grebenyuk. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for NukeExtensions.
-FOUNDATION_EXPORT double NukeExtensionsVersionNumber;
-
-//! Project version string for NukeExtensions.
-FOUNDATION_EXPORT const unsigned char NukeExtensionsVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <NukeExtensions/PublicHeader.h>
-
-
