+++ Tests/NukeTests/ImageDecoderRegistryTests.swift
