+++ Tests/NukeTests/ImagePipelineTests/ImagePipelineProgressiveDecodingTests.swift
@@ -187,82 +187,6 @@ class ImagePipelineProgressiveDecodingTests: XCTestCase {
         wait()
     }
 
-#if os(iOS) || os(tvOS) || os(macOS)
-
-    @MainActor
-    func testParitalImagesAreDisplayed() {
-        // Given
-        ImagePipeline.pushShared(pipeline)
-
-        let imageView = _ImageView()
-
-        let expectPartialImageProduced = self.expectation(description: "Partial Image Produced")
-        // We expect two partial images (at 5 scans, and 9 scans marks).
-        expectPartialImageProduced.expectedFulfillmentCount = 2
-
-        let expectedFinalLoaded = self.expectation(description: "Final Image Produced")
-
-        // When/Then
-        Nuke.loadImage(
-            with: Test.request,
-            into: imageView,
-            progress: { response, _, _ in
-                if let image = response?.image {
-                    XCTAssertTrue(imageView.image === image)
-                    expectPartialImageProduced.fulfill()
-                    self.dataLoader.resume()
-                }
-            },
-            completion: { result in
-                XCTAssertTrue(imageView.image === result.value?.image)
-                expectedFinalLoaded.fulfill()
-            }
-        )
-        wait()
-
-        ImagePipeline.popShared()
-    }
-
-    @MainActor
-    func testDisablingProgressiveRendering() {
-        // Given
-        ImagePipeline.pushShared(pipeline)
-
-        let imageView = _ImageView()
-
-        var options = ImageLoadingOptions()
-        options.isProgressiveRenderingEnabled = false
-
-        let expectPartialImageProduced = self.expectation(description: "Partial Image Produced")
-        // We expect two partial images (at 5 scans, and 9 scans marks).
-        expectPartialImageProduced.expectedFulfillmentCount = 2
-
-        let expectedFinalLoaded = self.expectation(description: "Final Image Produced")
-
-        // When/Then
-        Nuke.loadImage(
-            with: Test.request,
-            options: options,
-            into: imageView,
-            progress: { response, _, _ in
-                if response?.image != nil {
-                    XCTAssertNil(imageView.image)
-                    expectPartialImageProduced.fulfill()
-                    self.dataLoader.resume()
-                }
-            },
-            completion: { result in
-                XCTAssertTrue(imageView.image === result.value?.image)
-                expectedFinalLoaded.fulfill()
-            }
-        )
-        wait()
-
-        ImagePipeline.popShared()
-    }
-
-#endif
-
     func testProgressiveDecodingDisabled() {
         // Given
         var configuration = pipeline.configuration
