+++ Tests/NukeTests/ImageProcessorsTests/AnonymousTests.swift
