+++ Tests/NukeTests/ImageProcessorsTests/CircleTests.swift
