+++ Tests/NukeTests/ImageProcessorsTests/ImageProcessorsProtocolExtensionsTests.swift
